reduced number of files
